refactor: LineDto 주석 삭제 및 Domain 패키지로 이동
